requirements.txt
# Dependencies for CLAI - Command Line AI Team
# Install with: pip install -r requirements.txt

# CLI Framework
click>=8.1.0

# AI Provider SDKs
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Configuration
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Rich Terminal Output
rich>=13.0.0

# Interactive Shell
prompt_toolkit>=3.0.0

# Development (optional)
# pytest>=8.0.0
# pytest-asyncio>=0.23.0
# black>=24.0.0
# ruff>=0.5.0
